ensure calls to project fns are after cwd is set
