+++ packages/cli/src/commands/destroy/scaffold/__tests__/scaffold.test.js
@@ -5,8 +5,8 @@ import fs from 'fs'
 import '../../../../lib/test'
 
 import { getPaths, getDefaultArgs } from '../../../../lib'
-import { yargsDefaults as defaults } from '../../../generate'
 import { files } from '../../../generate/scaffold/scaffold'
+import { yargsDefaults as defaults } from '../../helpers'
 import { tasks } from '../scaffold'
 
 jest.mock('fs')
