+++ packages/cli/src/commands/generate/scaffold/__tests__/editableColumns.test.js
@@ -5,7 +5,7 @@ import path from 'path'
 import '../../../../lib/test'
 
 import { getDefaultArgs } from '../../../../lib'
-import { yargsDefaults as defaults } from '../../../generate'
+import { yargsDefaults as defaults } from '../../helpers'
 import * as scaffold from '../scaffold'
 
 describe('editable columns', () => {
